requirements.txt
PyMuPDF>=1.25.0
python-docx>=1.1.0
Jinja2>=3.1.0
Markdown>=3.6
jsonschema>=4.22.0
pytest>=8.0.0
pytest-cov>=5.0.0
